+++ src/dataStructures/graph/networkDelayTime.ts
@@ -1,16 +1,17 @@
 import { dijkstra } from './utils';
 
-function buildGraph(times: Array<[number, number, number]>, n: number): number[][] {
-  const graph: number[][] = Array(n)
-    .fill(null)
-    .map(() => Array(n)
-      .fill(Infinity));
+function buildGraph(times: Array<[number, number, number]>, n: number): Array<Array<[number, number]>> {
+  const graph: Array<Array<[number, number]>> = Array(n)
+    .fill(null);
   for (let i = 0; i < n; i++) {
-    graph[i][i] = 0;
+    graph[i] = [];
   }
 
   for (let i = 0; i < times.length; i++) {
-    graph[times[i][0] - 1][times[i][1] - 1] = times[i][2];
+    graph[times[i][0] - 1].push([
+      times[i][1] - 1,
+      times[i][2],
+    ]);
   }
 
   return graph;